move formatDifficulty to generic Common move suggestTarget command to bitcoin Cleanup include path
