+++ src/bytom/StratumSessionBytom.cc
@@ -28,7 +28,7 @@
 #include "bytom/bh_shared.h"
 
 #ifndef NO_CUDA
-#include "bytom/cutil/src/GpuTs.h"
+#include "cutil/src/GpuTs.h"
 #endif  //NO_CUDA
 
 /////////////////////////////StratumSessionBytom////////////////////////////
